requirements.txt
asgiref==3.5.2
Django==3.2.14
pytz==2022.1
sqlparse==0.4.2
asgiref==3.5.2
Django==3.2.14
djangorestframework==3.13.1
pytz==2022.1
rest-framework-generic-relations==2.1.0
sqlparse==0.4.2
